Silence warnings about worker signal handling
The current version of resque has some signal handling oddities
that are being dealt with in a transitional way:

http://hone.heroku.com/resque/2012/08/21/resque-signals.html

Setting this flag tells the Worker class to use the new signal
handling and silences a warning to the logs on each worker boot.

This doesn't actually affect the way we're managing workers, though,
since we don't spawn a new process per job. Workers don't actually
manage child processes at all in this system.