Workaround truncated listener/worker proclines
Under Linux, max custom procline length can in some circumstances be
limited to the original size of argv provided to execve(2). Because
the listener process is exec'd with most information being passed in
environ and a fairly short argv (can be "ruby bin/resqued listener"
in extreme cases), the listener and worker custom proclines end up
being truncated like follows:

    $ ps axo args | grep resqu[e]
    resqued-0.7.12 master [gen 1] [1 running] config/resqued/github-environment.rb config/resqued/github-staff.rb
    resqued-0.7.12 listener #1 4/4/4 [56d8c1a] [running] config/r
    resque-1.20.0: [56d8c1a] Waiting for lab_archive,lab_audit_lo
    resque-1.20.0: [56d8c1a] Waiting for lab_archive,lab_audit_lo
    resque-1.20.0: [56d8c1a] Waiting for lab_archive,lab_audit_lo
    resque-1.20.0: [56d8c1a] Waiting for lab_archive,lab_audit_lo

This patch adds an extra argument to the listener argv to pad out
the maximum possible custom procline size. It's not a great solution
but is the most reliable I've found. With it, I get full proclines
as you'd expect.