Try to reproduce a ECONNRESET between master and listener
The socket that's making ECONNRESET is the `UNIXSocket.pair` between the
master and listener.  If the master does `sock.close` while there's data
to read from the listener, and the listener then does `sock.read`, the
listener should get ECONNRESET. This commit adds the following behavior
that should let me tickle this:

`kill -USR1 $masterpid` will make the master close any listener socket
that has data available to read.

`kill -USR1 $listenerpid` will make the listener write something that
the master will ignore. When this happens after signalling the master
with USR1, the listener should end up getting ECONNRESET.

based on http://stackoverflow.com/questions/2974021/what-does-econnreset-mean-in-the-context-of-an-af-local-socket