docs: Change the master process's behavior on QUIT
Our current deploy strategy wants the master process to exit quickly.
The problem that led to #44 is what happens if the master process takes
more than 10s to exit. This branch will make the master branch exit
immediately on QUIT and TERM, since it doesn't need to coordinate between
the new and old listeners.