hash queue key to limit its size to 40 chars
A resque worker can process a large number of queues. If there are many
many workers with large queues, then the `RESQUED_STATE` environment
variable can become very large as well. The environment variable can
becomes so large that it exceeds the maximum length of arguments for a
new process. This will cause an E2BIG error.

Fix this problem by hashing the queue key to limits its size to 40 chars.