Implement --exec-on-hup
If resqued is started with --exec-on-hup, then when it receives a SIGHUP
it'll re-exec the master process before replacing the current listener.
It does the re-exec immediately, which has the effect of possibly losing
some data if there's a race between SIGHUP and another signal or IO.
Also, if the now-current version of resqued can't replace this one (e.g.
if resqued was downgraded), the master will crash. At this point, resqued
will rely on an external process monitor to restart it.